Rename article2 and purge article1
