+++ desktop/apps/article/client/article.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from 'desktop/apps/article2/components/App'
+import App from 'desktop/apps/article/components/App'
 import { rehydrateClient } from 'desktop/components/react/utils/renderReactLayout'
 
 export const init = () => {
