+++ desktop/apps/article/components/InfiniteScrollArticle.js
@@ -2,7 +2,7 @@ import * as React from 'react'
 import _ from 'underscore'
 import PropTypes from 'prop-types'
 import Waypoint from 'react-waypoint'
-import articlesQuery from 'desktop/apps/article2/queries/articles.js'
+import articlesQuery from 'desktop/apps/article/queries/articles.js'
 import colors from '@artsy/reaction-force/dist/Assets/Colors'
 import positronql from 'desktop/lib/positronql.coffee'
 import styled from 'styled-components'
