+++ desktop/apps/article/components/__tests__/EditButton.test.js
@@ -1,5 +1,5 @@
 import 'jsdom-global/register'
-import EditButton, { __RewireAPI__ as RewireApi } from 'desktop/apps/article2/components/EditButton'
+import EditButton, { __RewireAPI__ as RewireApi } from 'desktop/apps/article/components/EditButton'
 import React from 'react'
 import sinon from 'sinon'
 import { mount } from 'enzyme'
