--- desktop/apps/article/stylesheets/index.styl
@@ -1,74 +1,0 @@
-@require '../../../components/article/stylesheets'
-@require '../../../components/articles_grid/stylesheets'
-@require '../../../components/tooltips'
-@require '../../../components/channel_nav'
-@require './promoted_content.styl'
-
-#articles-show
-  margin-top 120px
-  &.is-loading
-    #article-loading
-      display block
-
-#articles-show.is-team-channel
-  margin-top 180px
-
-#article-loading
-  display none
-  margin-bottom 100px
-  .loading-spinner
-    position relative
-
-.body-welcome-header #articles-show
-  margin-top 180px
-
-#articles-body-container
-  position relative
-
-#articles-footer
-  @media screen and (max-width 550px)
-    margin 0 20px
-
-.articles-featured-articles
-  margin gutter-size auto
-  padding 20px 0
-  border-bottom 1px solid gray-lighter-color
-  max-width 1100px
-  h4
-    avant-garde()
-    font-size 13px
-    border-bottom 1px solid gray-lighter-color
-    padding-bottom 5px
-    margin-bottom 20px
-  li
-    border-right 1px solid gray-lighter-color
-    &:last-child
-      border 0
-    @media screen and (max-width 900px)
-      &.grid-item
-        width 100%
-        border-right 0
-        border-bottom 1px solid gray-lighter-color
-        padding 20px 0
-        &:first-child
-          padding-top 0
-        &:last-child
-          border 0
-          padding-bottom 0
-    img
-      width 70px
-      height 70px
-      margin-right 20px
-    p
-      font-size 20px
-      line-height 1.2em
-      width calc(100% \- 90px)
-      @media screen and (max-width 550px)
-        garamond-size('l-caption')
-    img, p
-      display inline-block
-      vertical-align middle
-  &.article-related-widget
-    margin-bottom 100px
-    @media screen and (max-width 550px)
-      margin-bottom 50px
