--- desktop/apps/article2/stylesheets/amp.styl
@@ -1,101 +1,0 @@
-@require '../../../components/stylus_lib'
-@require '../../../components/main_layout/stylesheets/embedded'
-@require '../../../components/main_layout/stylesheets/faux_underline'
-@require '../../../components/main_layout/stylesheets/icons'
-@require '../../../components/main_layout/stylesheets/typography'
-@require '../../../components/artwork_item/stylesheets'
-@require '../../../components/article/stylesheets/content'
-
-- amp-header-height = 54px
-
-global-reset()
-
-.body-amp
-  garamond()
-  font-size default-garamond-font-size
-  line-height 1.4em
-  text-rendering optimizeLegibility
-  min-width inherit
-  a
-    color black
-    cursor pointer
-
-  .large-garamond-header
-    garamond()
-    text-align left
-    display block
-    font-size 30px
-
-  figcaption, figcaption a
-    garamond-size('m-caption')
-    color gray-darker-color
-
-  .article-container, amp-article__logo-header
-    max-width 1250px
-    margin auto
-
-  .amp-article
-    &__logo-header
-      height amp-header-height
-      border-bottom 1px solid gray-lighter-color
-      display flex
-      align-items center
-      a
-        margin-left 20px
-
-    &__header
-      margin-top 50px
-
-    &__video-container
-      amp-video video
-        object-fit cover
-
-    &__artwork, &__image
-      margin-bottom 20px
-      figcaption
-        margin 10px 20px
-        a
-          text-decoration none
-
-    &__social
-      width 100vw
-      position fixed
-      bottom 0
-      left 0
-      margin 0
-      padding 0
-      z-index 1000
-      display flex
-      height 50px
-
-      amp-social-share
-        background-color transparent
-        background-image none
-
-      .share-item
-        width 33.333%
-        display flex
-        justify-content center
-        align-items center
-        &:hover
-          color white
-        &:first-child
-          border-right none
-
-    &__carousel
-      amp-img
-        height 350px
-        img
-          object-fit contain
-          object-position bottom
-        figcaption
-          height 150px
-
-    &__callout
-      a amp-img
-        min-width 120px
-
-  footer
-    margin-top 80px
-
-@require '../../../components/main_layout/footer'
