--- desktop/apps/article2/templates/amp_article.jade
@@ -1,37 +1,0 @@
-include amp_mixins
-include ../../../components/article/templates/mixins
-
-<!doctype html>
-html(amp)
-  head
-    +amp_head
-  body.body-article.body-amp
-    +amp_analytics
-    .amp-article__logo-header: a.icon-logotype(href="#{sd.APP_URL}")
-    if article.get('hero_section')
-      +amp_hero_section(article.get('hero_section'))
-    .article-container: .article-content
-      .amp-article__header.article-header.responsive-layout-container
-        h1.large-garamond-header= article.get('title')
-        if article.strip('lead_paragraph')
-          .article-lead-paragraph!= article.get('lead_paragraph')
-        +author-date
-        .amp-article__social.article-social.fixed
-          .share-item: amp-social-share.icon-facebook-f(type="facebook" width="30" height="30")
-          .share-item: amp-social-share.icon-twitter(type="twitter" width="30" height="30")
-          .share-item: amp-social-share.icon-mail(type="email" width="30" height="30")
-      for section in article.get('sections')
-        case section.type
-          when 'text'
-            +amp_text(section)
-          when 'image_collection'
-            +amp_image_collection(section)
-          when 'video'
-            +amp_video(section)
-          when 'embed'
-            +amp_embed(section)
-          when 'callout'
-            +amp_callout(section)
-          when 'image_set'
-            +amp_image_set(section)
-    include ../../../components/main_layout/footer/mobile_footer_menu
