--- desktop/apps/article2/templates/article.jade
@@ -1,27 +1,0 @@
-extends ../../../components/fair_layout/templates/conditional
-
-block head
-  include classic_meta
-
-append locals
-  - assetPackage = 'article2'
-  - bodyClass = bodyClass + ' ' + article.getBodyClass({})
-
-block body
-
-  //- Render channel nav if article is in a team channel
-  - var showClass = ''
-  if channel && channel.isTeam()
-    - showClass = 'is-team-channel'
-    include ../../../components/channel_nav/index
-
-  #articles-show(class=showClass)
-
-    //- Article body
-    #articles-body-container(role='article')
-      include ../../../components/article/templates/index
-
-  section#articles-footer.responsive-layout-container
-    //- Rendered on client
-
-  include ../../../components/main_layout/templates/json_ld
