--- desktop/apps/article2/templates/meta.jade
@@ -1,84 +1,0 @@
-- title = article.search_title || article.thumbnail_title || ''
-- url = sd.APP_URL + '/article/' + article.slug
-- search_title = article.search_title || article.thumbnail_title || ''
-- social_title = article.social_title || article.thumbnail_title || ''
-
-- search_description = article.search_description || article.description
-- social_description = article.social_description || article.description
-
-- social_image = article.social_image || article.thumbnail_image
-- keywords = []
-
-title= title
-meta( name='description', content= search_description )
-link( rel='canonical' href=url )
-if article.featured && article.published
-  link(rel='amphtml' href=(sd.APP_URL + '/article/' + article.slug + '/amp'))
-
-//- OG
-meta( property='og:title', content= social_title )
-meta( property='fb:app_id', content=sd.FACEBOOK_ID )
-meta( property='og:site_name' content='Artsy' )
-meta( property='og:url', content=url )
-meta( property='og:type', content='article' )
-meta( property='og:published_time', content=article.published_at )
-meta( property='article:published_time', content=article.published_at )
-meta( property='article:publisher', content='https://www.facebook.com/artsy/')
-meta( property='og:description', content= social_description )
-meta( property='og:image', content=social_image )
-meta( name='og:title' content=social_title)
-
-//- Twitter
-meta( name='twitter:site', content='@artsy' )
-meta( property='twitter:card', content='summary_large_image' )
-meta( property='twitter:url', content=url )
-meta( property='twitter:description', content= social_description )
-meta( property='twitter:image' content=social_image)
-meta( name='twitter:title' content=social_title)
-
-//- Keywords
-if article.keywords && article.keywords.length
-  - keywords = article.keywords.join(', ')
-  meta( property='news_keywords', content=keywords )
-  meta( name='keywords', content=keywords )
-  meta( property='article:tag', content=keywords )
-
-//- Sailthru Metatags
-if article.published && article.email_metadata
-  meta( name='sailthru.date' content=article.published_at)
-  - email = article.email_metadata
-  if email.headline
-    meta( name='sailthru.title', content=email.headline)
-  if email.author
-    meta( name='sailthru.author' content=email.author)
-  if email.credit_url
-    meta( name='sailthru.credit_url' content=email.credit_url)
-  if email.credit_line
-    meta( name='sailthru.credit_line' content=email.credit_line)
-  if email.image_url
-    - var src = email.image_url
-    meta( name='sailthru.image.full' content=crop(src, { width: 1200, height: 706 } ))
-    meta( name='sailthru.image.thumb' content=crop(src, { width: 900, height: 530 } ))
-  - var sailthru_keywords = ['article']
-  if article.featured === true
-    - sailthru_keywords = sailthru_keywords.concat(['magazine'])
-  if keywords.length
-    - sailthru_keywords = sailthru_keywords.concat(keywords)
-  meta( name='sailthru.tags' content=sailthru_keywords.join(','))
-
-//- Author Metatags
-if article.contributing_authors && article.contributing_authors.length
-  - for author in article.contributing_authors
-      if author.name
-        meta( name='author', content=author.name )
-
-//- Unica Font
-link( rel='stylesheet' type='text/css' href="#{sd.WEBFONT_URL + '/unica-webfonts.css'}")
-
-//- No Index
-if !article.indexable
-  meta( name='robots', content='noindex' )
-
-//- Slick styles
-<link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick.min.css" />
-<link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick-theme.min.css"/>
