+++ desktop/assets/article.styl
@@ -1,2 +1,3 @@
-@require '../apps/article/stylesheets'
+@require '../apps/article/stylesheets/classic'
 @require '../components/venice_banner'
+@require '../components/merry_go_round'
