--- desktop/assets/article2.coffee
@@ -1,13 +1,0 @@
-require('backbone').$ = $
-sd = require('sharify').data
-
-layout = sd.ARTICLE?.layout
-
-$ ->
-  if location.pathname.match('article')
-    if layout is 'classic'
-      { init } = require('../apps/article2/client/classic.js')
-      init()
-    else
-      { init } = require('../apps/article2/client/article.js')
-      init()
