--- desktop/assets/article2.styl
@@ -1,3 +1,0 @@
-@require '../apps/article2/stylesheets/classic'
-@require '../components/venice_banner'
-@require '../components/merry_go_round'
